+++ .idea/encodings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/comp0034-cw1/data/SeoulBikeData.csv" charset="windows-1252" />
+    <file url="file://$PROJECT_DIR$/seoulbike_app/data/SeoulBikeData.csv" charset="windows-1252" />
   </component>
 </project>
