+++ README.md
@@ -5,9 +5,14 @@ This is Soyeong Bak's 2022/2023 COMP0034 Coursework 1.
 
 # Set-up instructions
 
-1. You can clone the repository or access it through: https://github.com/sy-122/comp0034-cw1-i-sy-122.git
-2. You need an active virtual environment. 
+1. You can git clone or donwload the repository through:
+```
+https://github.com/sy-122/comp0034-cw1-i-sy-122.git
+```
+
+2. Then you need an active virtual environment.
 3. Install the requirements from requirements.txt.
+4. Run the [dash_app.py](seoulbike_app/dash_app.py)
 
 # Visualisation design
 
@@ -24,7 +29,7 @@ Hence, although there is a navigation bar with two different pages those pages a
 currently not active but will be in the future. 
 
 The dash app has two tabs that divide the visualizations as the first one uses time related variables and the 
-second tab uses other variables, which are related to the weather. There are different functions
-in the app that allows uses to take more control of the dash app making it more interactive.
+second tab uses other variables, which are related to the weather. There are different interactive functions
+in the app that allows users to take more control of the dash app.
 
-*it can take time for the visualization to load so please be patient*
+*It can take time for the visualization to load so please be patient.*
